新增初版冗余点去除功能；平滑操作还需修改；细节优化。 1. 新增冗余点去除功能。实现方法为：若两个节点扩展的方向一致，则舍弃前一个。 2. 贝塞尔曲线平滑现在使用4个控制点一组进行平滑，点的密度激增，t_step为0.01，后续考虑降采样和实现分段贝塞尔曲线的效果。 3. 所有float类型替换成double类型。 4. 添加resetMap函数，用于每次规划完成后重置地图数据。

+++ src/my_astar/include/my_astar/astar.h
@@ -27,11 +27,11 @@ class Astar : public GlobalPlannerInterface
         };
 
         cv::Point2i point;
-        float cost = 0;
+        double cost = 0;
 
-        float g = 0;
-        float h = 0;
-        float f = 0;
+        double g = 0;
+        double h = 0;
+        double f = 0;
         NodeType type = NodeType::UNKNOWN;
         Node * parent_node = nullptr;
 
@@ -55,7 +55,7 @@ public:
     bool setEndPoint(const cv::Point2i p) override;
 
     bool getRawPath(std::vector<cv::Point2i> & path) override;
-    bool getSmoothPath(std::vector<cv::Point2f> & path) override;
+    bool getSmoothPath(std::vector<cv::Point2d> & path) override;
 
 private:
     std::vector<std::vector<Node>> map_;
@@ -64,5 +64,5 @@ private:
 
     HeuristicsType type_;
 
-    float getH(cv::Point2i p);
+    double getH(cv::Point2i p);
 };
