+++ src/my_astar/include/my_astar/bezier_curve.h
@@ -15,7 +15,7 @@ public:
     /// @param bezier_points 输出的平滑后的稠密序列点
     /// @param t_step 贝塞尔平滑中t的步增长度，t∈[0, 1]，每一步生成一个输出点
     template <typename PointType>
-    void smooth_curve(const std::vector<PointType> & control_points, std::vector<cv::Point2f> & bezier_points, double t_step = 0.01)
+    void smooth_curve(const std::vector<PointType> & control_points, std::vector<cv::Point2d> & bezier_points, double t_step = 0.01)
     {
         if (control_points.size() == 0)
         {
@@ -25,7 +25,7 @@ public:
         if (control_points.size() == 1)
         {
             bezier_points.clear();
-            bezier_points.push_back(cv::Point2f(control_points[0].x, control_points[0].y));
+            bezier_points.push_back(cv::Point2d(control_points[0].x, control_points[0].y));
             return;
         }
         
@@ -33,7 +33,7 @@ public:
         size_t n = control_points.size() - 1;
         for (double t = 0; t <= 1; t += t_step)
         {
-            cv::Point2f pt(0.0f, 0.0f);
+            cv::Point2d pt(0.0f, 0.0f);
             for (size_t i = 0; i <= n; i++)
             {
                 pt.x += (C_n_r(n, i) * std::pow(t, i) * std::pow(1 - t, n - i) * control_points[i].x);
